pyproject.toml
[tool.poetry]
name = "Diary of visits to restaurants"
version = "0.1.0"
description = "Assigment to ThinkEasy"
authors = ["Artem Jasan <artjasan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
dj-rest-auth = "^2.2.4"
django-allauth = "^0.50.0"
djangorestframework-simplejwt = "^5.1.0"
psycopg2 = "^2.9.3"
flake8 = "^4.0.1"
mypy = "^0.950"
isort = "^5.10.1"
black = "^22.3.0"
pre-commit = "^2.19.0"
requests = "^2.27.1"
drf-spectacular = "^0.22.1"
pytest = "^7.1.2"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.3.1"
responses = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
pythonpath = [".", "restaurant_visits_project"]

[tool.isort]
src_paths = "restaurant_visits_diary"
known_local_folder = "restaurant_visits_diary/"
line_length  = 120
profile = "black"

[tool.mypy]
files = "restaurant_visits_diary"
ignore_missing_imports = true
exclude = """(?x)^(
        .*/migrations/.* |
        .*/tests/.*
    )$"""

[tool.black]
line_length = 120
include = '\.pyi?$'
exclude = '''
/(
  | \.venv
  | migrations
)/
'''
